+++ src/Factory/PurifierFilterFactory.php
@@ -9,13 +9,17 @@ namespace Soflomo\Purifier\Factory;
 
 use Soflomo\Purifier\PurifierFilter;
 use Zend\Filter\FilterPluginManager;
+use Zend\ServiceManager\MutableCreationOptionsInterface;
+use Zend\ServiceManager\MutableCreationOptionsTrait;
 
-class PurifierFilterFactory
+class PurifierFilterFactory implements MutableCreationOptionsInterface
 {
+    use MutableCreationOptionsTrait;
+
     public function __invoke(FilterPluginManager $filterPluginManager)
     {
         $htmlPurifier = $filterPluginManager->getServiceLocator()->get('HTMLPurifier');
 
-        return new PurifierFilter($htmlPurifier);
+        return new PurifierFilter($htmlPurifier, $this->getCreationOptions());
     }
 }